Demo profile creation

+++ app/src/main/java/com/company/metrix/model/Employee.kt
@@ -3,7 +3,7 @@ package com.company.metrix.model
 data class Employee(
     val id: Int,
     val team: Int,
-    val ratings: List<Float>,
+    val ratings: List<Double>,
     val strongSkills: List<String>,
     val weakSkills : List<String>,
     val achievements: List<String>
